+++ Changelog.md
@@ -3,6 +3,7 @@
 
 ## Added
     * New filter to result_set
+    * New new copy and settings buttons for result set
 ## [0.5.9] - 03.18.2019
 ## Added
     * New componens: status filter. It is status with count of elements